builderz/backend/requirements.txt
fastapi==0.110.0
uvicorn==0.30.0
rasterio==1.4.3
numpy>=1.24,<2.0.0
pandas==2.2.3
matplotlib==3.7.3
pillow==11.1.0
imageio==2.33.1
langchain==0.3.19
langchain-groq==0.2.4
langgraph==0.3.2
pydantic>=2.7.4,<3.0.0
python-dotenv==1.0.1 